fix: remove unused import and move type definitions
