+++ src/web/app/types/auth/index.d.ts
@@ -3,3 +3,14 @@ export interface DecodedToken {
   aud: string;
   identity_proofing_level: string;
 }
+
+declare module "next-auth" {
+  interface User {
+    firstName?: string;
+    lastName?: string;
+    dob?: string;
+    nhsNumber?: string;
+    identityLevel?: string;
+    accessToken?: string;
+  }
+}
